refactor components, update keycaps page with images
