+++ pages/_app.js
@@ -1,4 +1,5 @@
 // import { PageLayout } from "@/layouts/PageLayout";
+import { CloudinaryContext, Transformation, Image } from "cloudinary-react";
 import { QueryClient, QueryClientProvider } from "react-query";
 import "../styles/globals.css";
 
@@ -6,10 +7,12 @@ const queryClient = new QueryClient();
 
 function MyApp({ Component, pageProps }) {
   return (
-    <QueryClientProvider client={queryClient}>
-      {/* <PageLayout></PageLayout> */}
-      <Component {...pageProps} />
-    </QueryClientProvider>
+    <CloudinaryContext cloudName="brianhan">
+      <QueryClientProvider client={queryClient}>
+        {/* <PageLayout></PageLayout> */}
+        <Component {...pageProps} />
+      </QueryClientProvider>
+    </CloudinaryContext>
   );
 }
