+++ pages/keyboards.js
@@ -1,8 +1,7 @@
 import { request, gql } from "graphql-request";
-import { CloudinaryContext, Transformation, Image } from "cloudinary-react";
-import { useMedia } from "use-media";
 
 import { PageLayout } from "@/components/PageLayout/PageLayout";
+import { GridGallery } from "@/components/GridGallery/GridGallery";
 import styles from "@/styles/Keyboards.module.css";
 
 const ENDPOINT = process.env.PROD_GRAPHQL_ENDPOINT;
@@ -25,47 +24,13 @@ export async function getStaticProps() {
   };
 }
 
-function GridImage({ name, publicId }) {
-  const isMobile = useMedia({ maxWidth: 750 });
-
-  return (
-    <div
-      className={styles.GridItem}
-      style={{
-        display: "flex",
-        justifyContent: "center",
-        alignItems: "center",
-      }}
-    >
-      <Image publicId={publicId} alt={`picture of a ${name} keyboard`}>
-        <Transformation width={isMobile ? "400" : "600"} crop="scale" />
-      </Image>
-
-      <span className={styles.KeyboardName}>
-        <p>{name}</p>
-      </span>
-    </div>
-  );
-}
-
 export default function Keyboards({ data }) {
   return (
-    <CloudinaryContext cloudName="brianhan">
-      <PageLayout className={styles.Keyboards}>
-        <div className={styles.Title}>
-          <h2>Keyboards</h2>
-        </div>
-
-        <div className={styles.Grid}>
-          {data.keyboards.map((keyboard) => (
-            <GridImage
-              name={keyboard.name}
-              key={keyboard.id}
-              publicId={keyboard.thumbnail_public_id}
-            />
-          ))}
-        </div>
-      </PageLayout>
-    </CloudinaryContext>
+    <PageLayout className={styles.Keyboards}>
+      <div className={styles.PageTitle}>
+        <h2>Keyboards</h2>
+      </div>
+      <GridGallery images={data.keyboards} />
+    </PageLayout>
   );
 }
